requirements.txt
flask
gunicorn
tensorflow
transformers
googletrans==4.0.0-rc1
pyenchant

